add handleStorage
